requirements.txt
pyttsx3
speechrecognition
wikipedia
pywhatkit
python-dotenv
groq
win10toast
pyaudio
mtranslate
tk
customtkinter
